feat: file handling plugin
